[REFACTOR] RoomService 일부 메서드 RoomMemberService로 이동 (#133)
